Remove constant from config
