+++ src/support/steps/ll-css-bg-image.ts
@@ -1,7 +1,6 @@
 import { When, Then } from '@cucumber/cucumber';
 import { ICustomWorld } from "../../common/custom-world";
 import { expect } from "@playwright/test";
-import {LL_BACKGROUND_IMAGES} from '../../../config/wp.config';
 
 Then('I must see the correct style in the head', async function (this: ICustomWorld) {
   const html = await this.page.evaluate(async () => {
@@ -60,6 +59,73 @@ Then('I must see the correct style in the head', async function (this: ICustomWo
 
   expect(isMatch, failMessage).toBeTruthy();
 });
+
+const LL_BACKGROUND_IMAGES = {
+    // eslint-disable-next-line @typescript-eslint/naming-convention
+    lazyload_css_background_images: {
+        initialImages: [
+            'https://new.rocketlabsqa.ovh/wp-content/rocket-test-data/images/test.png',
+            'https://e2e.rocketlabsqa.ovh/wp-content/rocket-test-data/images/image-insidescript.jpeg',
+            'https://new.rocketlabsqa.ovh/wp-content/rocket-test-data/images/paper.jpeg',
+            'https://new.rocketlabsqa.ovh/wp-content/rocket-test-data/images/test_inline1.jpeg',
+            'https://e2e.rocketlabsqa.ovh/wp-content/rocket-test-data/images/test_inline2.jpeg',
+            'https://e2e.rocketlabsqa.ovh/wp-content/rocket-test-data/images/test_internal0.webp'
+        ],
+        lazyLoadedImages: [
+            'https://new.rocketlabsqa.ovh/wp-content/rocket-test-data/image/test3.webp',
+            'https://e2e.rocketlabsqa.ovh/wp-content/rocket-test-data/image/file_example_JPG_100kB.jpg',
+            'https://rocketlabsqa.ovh/wp-content/rocket-test-data/images/Spain.PNG',
+            'https://e2e.rocketlabsqa.ovh/wp-content/rocket-test-data/images/paper.jpeg',
+            'https://e2e.rocketlabsqa.ovh/wp-content/rocket-test-data/images/test_internal4.jpg',
+            'https://e2e.rocketlabsqa.ovh/wp-content/rocket-test-data/images/testnotExist.png',
+            'https://e2e.rocketlabsqa.ovh/wp-content/rocket-test-data/images/wp-rocket.svg',
+            'https://new.rocketlabsqa.ovh/wp-content/rocket-test-data/images/butterfly.avif',
+            'https://e2e.rocketlabsqa.ovh/wp-content/rocket-test-data/images/file_example_TIFF_1MB.tiff',
+            'https://fastly.picsum.photos/id/976/200/300.jpg?hmac=s1Uz9fgJv32r8elfaIYn7pLpQXps7X9oYNwC5XirhO8',
+            'https://rocketlabsqa.ovh/wp-content/rocket-test-data/images/fixtheissue.jpg',
+            'https://mega.wp-rocket.me/avada/wp-content/rocket-test-data/prague-conference-center-1056491.jpg',
+            'https://e2e.rocketlabsqa.ovh/wp-content/rocket-test-data/images/butterfly%202.avif',
+            'https://new.rocketlabsqa.ovh/wp-content/rocket-test-data/images/paper%C3%A9quipesTest.jpeg',
+            'https://new.rocketlabsqa.ovh/wp-content/rocket-test-data/images/test_internal2.jpg',
+            'https://e2e.rocketlabsqa.ovh/wp-content/plugins/revslidertest1/public/assets/assets/test_internal3.jpg',
+            'https://e2e.rocketlabsqa.ovh/wp-content/rocket-test-data/images/test.png',
+            'https://e2e.rocketlabsqa.ovh/wp-content/rocket-test-data/image/test3.gif',
+            'https://e2e.rocketlabsqa.ovh/wp-content/rocket-test-data/images/test_external1.jpeg',
+            'https://e2e.rocketlabsqa.ovh/test.png',
+            'https://upload.wikimedia.org/wikipedia/commons/1/11/Test-Logo.svg',
+            'https://e2e.rocketlabsqa.ovh/kot%C5%82.png',
+            'https://new.rocketlabsqa.ovh/wp-content/rocket-test-data/styles/assets/images/relative1.jpeg',
+            'https://new.rocketlabsqa.ovh/wp-content/rocket-test-data/styles/assets/images/relative2.jpg',
+            'https://e2e.rocketlabsqa.ovh/wp-content/rocket-test-data/images/lcp/testsvg.svg',
+            'https://new.rocketlabsqa.ovh//wp-content/rocket-test-data/images/wp-rocket.svg',
+            'https://new.rocketlabsqa.ovh/wp-content/rocket-test-data/images/wp-rocket2.svg',
+            'https://e2e.rocketlabsqa.ovh/new.rocketlabsqa.ovh/wp-content/rocket-test-data/images/test_internal2.jpg',
+            'https://new.rocketlabsqa.ovh/wp-content/rocket-test-data/images/nature.jpeg'
+        ]
+    },
+    // eslint-disable-next-line @typescript-eslint/naming-convention
+    ll_bg_css_single_colon: {
+        initialImages: [
+            'https://e2e.rocketlabsqa.ovh/wp-content/rocket-test-data/images/fabio-sasso-UgpCjt4XLTY-unsplash.jpg',
+            'https://e2e.rocketlabsqa.ovh/wp-content/rocket-test-data/images/underline.png',
+            'https://e2e.rocketlabsqa.ovh/wp-content/rocket-test-data/images/maxime-lebrun-6g3Akg708E0-unsplash.jpg'
+        ],
+        lazyLoadedImages: [
+            'https://e2e.rocketlabsqa.ovh/wp-content/rocket-test-data/images/miguel-luis-6wxFtwSuXHQ-unsplash.jpg'
+        ]
+    },
+    // eslint-disable-next-line @typescript-eslint/naming-convention
+    ll_bg_css_double_colon:{
+        initialImages: [
+            'https://e2e.rocketlabsqa.ovh/wp-content/rocket-test-data/images/fabio-sasso-UgpCjt4XLTY-unsplash.jpg',
+            'https://e2e.rocketlabsqa.ovh/wp-content/rocket-test-data/images/underline.png',
+            'https://e2e.rocketlabsqa.ovh/wp-content/rocket-test-data/images/maxime-lebrun-6g3Akg708E0-unsplash.jpg'
+        ],
+        lazyLoadedImages: [
+            'https://e2e.rocketlabsqa.ovh/wp-content/rocket-test-data/images/miguel-luis-6wxFtwSuXHQ-unsplash.jpg',
+        ]
+    }
+};
 
 When('I go to {string} check initial image loaded', async function (this: ICustomWorld, page) {
     const images = [];
